requirements.txt
altair==4.2.2
anyio==3.6.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dill==0.3.6
dnspython==2.3.0
entrypoints==0.4
evidently==0.2.8
fastapi==0.95.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.4.1
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-boto3-s3==1.26.104
neuro-mf==0.0.5
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.5.0
pip-chill==1.0.1
pkgutil_resolve_name==1.3.10
plotly==5.14.1
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.0
pymongo==4.3.3
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rich==13.3.4
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
statsmodels==0.13.5
streamlit==1.22.0
tenacity==8.2.2
threadpoolctl==3.1.0
tld==0.13
tldextract==3.4.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
types-awscrt==0.16.13.post1
types-s3transfer==0.6.0.post7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.1
wincertstore==0.2
whois==0.9.27
zipp==3.15.0
